refactor: move nested app route components to /components directory (#113)
Move maintenance components from app/admin/maintenance/*/components/
to the centralized components/admin/maintenance/ directory for better
discoverability and consistency with project structure guidelines.
